+++ package.json
@@ -47,6 +47,7 @@
     "babel-plugin-transform-object-rest-spread": "6.26.0",
     "babel-preset-es2015": "6.22.0",
     "babel-preset-react": "6.22.0",
+    "bowser": "1.9.4",
     "cheerio": "1.0.0-rc.2",
     "classnames": "2.2.5",
     "cookie": "0.2.2",
